submission-pemda/requirements.txt
requests==2.32.3
beautifulsoup4==4.12.3
pandas==2.2.3
gspread==6.1.2
oauth2client==4.1.3
psycopg2-binary==2.9.9
pytest==8.3.3
pytest-cov==5.0.0